Deleted the multiselectspinner. Created another better alternative.
Uses alert-dialog so the user can choose what courses and education the user has.
Enables the user to get (maxiumum) 2 extra points for education as military and "folkehøgskole".
Returns and array with selected courses and education.
Changed the color of the numberpicker.
Not calculating score after what kind of courses the user has, so this must be implemented.