Worked on basic GUI-testing functionality in Android Studio using Espresso testing framework.
Created a test that checks for objects on the log in page. Struggled a lot with different errors, but eventually managed to find a workaround using injectInstrumentation. Tests seems to be working; at least they run and complete without errors. Would be nice to know if they _actually_ tested anything...