Changed the CHATBOT-interface and added a new function.
Added 3 textview-buttons on the top of the screen (that changes color when pressed). If you press the "back"-button, you are directed to the menu page. If you press the "help"-button, then an alertdialog pops out with help-instructions. If you press the "uniBOT"-button, sick things will happen: a random question will appear from uniBOT to the user in the chat-dialog. This is done by an arraylist with different questions (for example; "do you want to know more about physics?") made in the Chatbot-class (this needs to get more questions), and if the user answers "yes", the coherent question (in another arraylist (in the physics-example; it will be "tell the user more about physics")) is sent to API.AI, so you get a maximum flow between the user and uniBOT.

Also changed the background color for the listview (so it is a small color-difference from where you type the text), and added bars under/over different textviews. The send-button also changes color when pressed.