Finised writing GUI-tests for the login screen and started working on tests for the About Unibot and About Us subpages.
The test supposed to check the "register user"-feature currently fail due to the fact the example-user is already in the database. The test checks if the "sigh out"-button exists after pressing register, at which point the "user" recieves a toast stating the account register didn't work. Need to add some type of "if"-statement here to avoid this happening and failing the test. Perhaps there is a way to check for toasts