Fixed; Sign_in: changed so the user has to wait 5 seconds in the sign_in screen (toast appears) so firebase can load the user's data. Fixed; Add_information: changed so the information from the alertdialogs contains the right values, and if "Matematikk R2" is deselected -> it won't appear when the user presses the settings-wheel.
Can't save the user's grades before Jørgen creates field for them in firebase.

Found a bug: if the user presses outside the alertdialog when (s)he is selecting R2-grade, then the user doesn't have to type it in. Should we let the user avoid setting the R2 grade?